+++ R/cat2cat.R
@@ -24,9 +24,9 @@
 #' `data`, `cat_var`, `method`, `features` and optional `args`.
 #' @details
 #' data args
-#' \itemize{
+#' \describe{
 #'  \item{"old"}{ data.frame older time point in a panel}
-#'  \item{"new"} { data.frame more recent time point in a panel}
+#'  \item{"new"}{ data.frame more recent time point in a panel}
 #'  \item{"time_var"}{ character(1) name of the time variable.}
 #'  \item{"cat_var"}{ character(1) name of the categorical variable.}
 #'  \item{"cat_var_old"}{
@@ -51,7 +51,7 @@
 #'  }
 #' }
 #' mappings args
-#' \itemize{
+#' \describe{
 #'  \item{"trans"}{ data.frame with 2 columns - mapping (transition) table -
 #'   all categories for cat_var in old and new datasets have to be included.
 #'   First column contains an old encoding and second a new one.
@@ -70,7 +70,7 @@
 #'  }
 #' }
 #' Optional ml args
-#' \itemize{
+#' \describe{
 #'  \item{"data"}{ data.frame - dataset with features and the `cat_var`.}
 #'  \item{"cat_var"}{ character(1) - the dependent variable name.}
 #'  \item{"method"}{
@@ -137,19 +137,26 @@
 #'   mappings = list(trans = trans2, direction = "forward")
 #' )
 #'
-#' # additional probabilities from knn
-#' occup_ml <- cat2cat(
-#'   data = list(
-#'     old = occup_old, new = occup_new, cat_var = "code", time_var = "year"
-#'   ),
-#'   mappings = list(trans = trans, direction = "backward"),
-#'   ml = list(
+#' mappings <- list(trans = trans, direction = "backward")
+#'
+#' ml_setup <- list(
 #'     data = occup_small[occup_small$year >= 2010, ],
 #'     cat_var = "code",
 #'     method = "knn",
 #'     features = c("age", "sex", "edu", "exp", "parttime", "salary"),
 #'     args = list(k = 10)
-#'   )
+#' )
+#'
+#' # ml model performance check
+#' print(cat2cat_ml_run(mappings, ml_setup))
+#'
+#' # additional probabilities from knn
+#' occup_ml <- cat2cat(
+#'   data = list(
+#'     old = occup_old, new = occup_new, cat_var = "code", time_var = "year"
+#'   ),
+#'   mappings = mappings,
+#'   ml = ml_setup
 #' )
 #' }
 #'
@@ -179,37 +186,13 @@ cat2cat <- function(data = list(
   stopifnot(is.list(mappings))
   stopifnot(is.list(ml))
 
-  # data validation
-  stopifnot(inherits(data$old, "data.frame"))
-  stopifnot(inherits(data$new, "data.frame"))
-  stopifnot(!is.null(data$cat_var) ||
-    (!is.null(data$cat_var_old) && !is.null(data$cat_var_new)))
+  # Backward compatibility
+  if (is.null(data$freqs_df)) data$freqs_df <- mappings$freqs_df
   if (is.null(data$cat_var_old)) data$cat_var_old <- data$cat_var
   if (is.null(data$cat_var_new)) data$cat_var_new <- data$cat_var
-  stopifnot(all(c(data$cat_var_old, data$time_var) %in% colnames(data$old)))
-  stopifnot(all(c(data$cat_var_new, data$time_var) %in% colnames(data$new)))
-  stopifnot(
-    is.null(data$multiplier_var) ||
-      (data$multiplier_var %in% colnames(data$new) &&
-        data$multiplier_var %in% colnames(data$old))
-  )
 
-  # For backward compatibility
-  if (is.null(data$freqs_df)) data$freqs_df <- mappings$freqs_df
-  stopifnot(is.null(data$freqs_df) ||
-    (is.data.frame(data$freqs_df) && ncol(data$freqs_df) == 2))
-  stopifnot((length(unique(data$old[[data$time_var]])) == 1) &&
-    (length(unique(data$new[[data$time_var]])) == 1))
-  stopifnot(is.null(data$id_var) || ((data$id_var %in% colnames(data$old)) &&
-    (data$id_var %in% colnames(data$new)) &&
-    !anyDuplicated(data$old[[data$id_var]]) &&
-    !anyDuplicated(data$new[[data$id_var]])))
-
-  # mappings validation
-  stopifnot(all(vapply(mappings, Negate(is.null), logical(1))))
-  stopifnot(all(c("trans", "direction") %in% names(mappings)))
-  stopifnot(isTRUE(mappings$direction %in% c("forward", "backward")))
-  stopifnot(is.data.frame(mappings$trans) && ncol(mappings$trans) == 2)
+  validate_data(data)
+  validate_mappings(mappings)
 
   mapps <- get_mappings(mappings$trans)
 
@@ -311,175 +294,8 @@ cat2cat <- function(data = list(
   res
 }
 
-#' The internal function used in the cat2cat one
-#' @description apply the ml models to the cat2cat data
-#' @param ml `list` the same `ml` argument as provided to `cat2cat` function.
-#' @param mapp `list` a mapping table
-#' @param target_data `data.frame`
-#' @param cat_var_target `character(1)` name of the categorical variable
-#' in the target period.
-#' @keywords internal
-cat2cat_ml <- function(ml, mapp, target_data, cat_var_target) {
-  stopifnot(all(c("method", "features") %in% names(ml)))
-  stopifnot(all(ml$features %in% colnames(target_data)))
-  stopifnot(all(ml$features %in% colnames(ml$data)))
-  stopifnot(all(vapply(
-    target_data[, ml$features, drop = FALSE],
-    function(x) is.numeric(x) || is.logical(x), logical(1)
-  )))
-  stopifnot(all(vapply(
-    ml$data[, ml$features, drop = FALSE],
-    function(x) is.numeric(x) || is.logical(x), logical(1)
-  )))
-  stopifnot(all(ml$method %in% c("knn", "rf", "lda")))
-  stopifnot(ml$cat_var %in% colnames(ml$data))
-  stopifnot(cat_var_target %in% colnames(target_data))
-
-  features <- unique(ml$features)
-  methods <- unique(ml$method)
-  ml_names <- paste0("wei_", methods, "_c2c")
-
-  target_data[, ml_names] <- target_data["wei_freq_c2c"]
-
-  cat_ml_year_g <- split(
-    ml$data[, c(features, ml$cat_var), drop = FALSE],
-    factor(ml$data[[ml$cat_var]], exclude = NULL)
-  )
-  target_data_cats <- target_data[[cat_var_target]]
-  target_data_cat_c2c <- split(
-    target_data,
-    factor(target_data_cats, exclude = NULL)
-  )
-
-  for (cat in unique(names(target_data_cat_c2c))) {
-    try(
-      {
-        matched_cat <- match(cat, names(target_data_cat_c2c))
-        target_data_cat <- target_data_cat_c2c[[matched_cat]]
-        dis <- do.call(rbind, cat_ml_year_g[mapp[[match(cat, names(mapp))]]])
-        udc <- unique(dis[[ml$cat_var]])
-        if (length(udc) <= 1) {
-          target_data_cat_c2c[[matched_cat]][ml_names] <-
-            target_data_cat$wei_freq_c2c
-          next
-        }
-        if (
-          length(unique(target_data_cat$g_new_c2c)) > 1 &&
-            length(udc) >= 1 &&
-            nrow(target_data_cat) > 0 &&
-            any(unique(target_data_cat$g_new_c2c) %in% names(cat_ml_year_g))
-        ) {
-          base_ml <-
-            target_data_cat[
-              !duplicated(target_data_cat[["index_c2c"]]),
-              c("index_c2c", features)
-            ]
-          cc <- complete.cases(base_ml[, features])
-          for (m in methods) {
-            ml_name <- paste0("wei_", m, "_c2c")
-            if (m == "knn") {
-              if (
-                suppressPackageStartupMessages(
-                  requireNamespace("caret", quietly = TRUE)
-                )
-              ) {
-                kkk <- suppressWarnings(
-                  caret::knn3(
-                    x = dis[, features, drop = FALSE],
-                    y = factor(dis[[ml$cat_var]]),
-                    k = min(ml$args$k, ceiling(nrow(dis) / 4))
-                  )
-                )
-                pp <- as.data.frame(
-                  stats::predict(
-                    kkk,
-                    base_ml[cc, features, drop = FALSE],
-                    type = "prob"
-                  )
-                )
-              } else {
-                stop(
-                  paste(
-                    "Please install caret package to use the knn model",
-                    "in the cat2cat function."
-                  )
-                )
-              }
-            } else if (m == "rf") {
-              if (
-                suppressPackageStartupMessages(
-                  requireNamespace("randomForest", quietly = TRUE)
-                )
-              ) {
-                kkk <- suppressWarnings(
-                  randomForest::randomForest(
-                    y = factor(dis[[ml$cat_var]]),
-                    x = dis[, features, drop = FALSE],
-                    ntree = min(ml$args$ntree, 100)
-                  )
-                )
-                pp <- as.data.frame(
-                  stats::predict(
-                    kkk,
-                    base_ml[cc, features, drop = FALSE],
-                    type = "prob"
-                  )
-                )
-              } else {
-                stop(paste(
-                  "Please install randomForest package to use",
-                  "the rf model in the cat2cat function."
-                ))
-              }
-            } else if (m == "lda") {
-              kkk <- suppressWarnings(
-                MASS::lda(
-                  grouping = factor(dis[[ml$cat_var]]),
-                  x = as.matrix(dis[, features, drop = FALSE])
-                )
-              )
-              pp <- as.data.frame(
-                stats::predict(
-                  kkk,
-                  as.matrix(base_ml[cc, features, drop = FALSE])
-                )$posterior
-              )
-            }
-            ll <- setdiff(unique(target_data_cat$g_new_c2c), colnames(pp))
-            # imputing rest of the class to zero prob
-            if (length(ll)) {
-              pp[ll] <- 0
-            }
-            pp_stack <- utils::stack(pp)
-            pp[["index_c2c"]] <- base_ml[["index_c2c"]][cc]
-            res <- cbind(pp_stack, index_c2c = rep(pp$index_c2c, ncol(pp) - 1))
-            colnames(res) <- c("val", "g_new_c2c", "index_c2c")
-            ress <- merge(
-              target_data_cat[, c("index_c2c", "g_new_c2c")],
-              res,
-              by = c("index_c2c", "g_new_c2c"),
-              all.x = TRUE,
-              sort = FALSE
-            )
-            resso <- ress[order(ress$index_c2c), ]
-            target_data_cat_c2c[[
-              match(cat, names(target_data_cat_c2c))
-            ]][[ml_name]] <- resso$val
-          }
-        }
-      },
-      silent = TRUE
-    )
-  }
-
-  target_data <- do.call(rbind, target_data_cat_c2c)
-  target_data <- target_data[order(target_data[["index_c2c"]]), ]
-
-  list(target_data = target_data)
-}
-
 #' Validate if the trans table contains all proper mappings
-#' @param cats_target vector of unique target period categories
+#' @param u_cats_target vector of unique target period categories
 #' @param mapp transition (mapping) table process with `get_mappings`,
 #'  the "to_base" direction is taken.
 #' @keywords internal
@@ -537,4 +353,39 @@ resolve_frequencies <- function(cat_base_year,
     )
   }
   freqs
+}
+
+#" Validate cat2cat data
+#' @keywords internal
+validate_data <- function(data) {
+  stopifnot(inherits(data$old, "data.frame"))
+  stopifnot(inherits(data$new, "data.frame"))
+  stopifnot(!is.null(data$cat_var_old) && !is.null(data$cat_var_new))
+
+  stopifnot(all(c(data$cat_var_old, data$time_var) %in% colnames(data$old)))
+  stopifnot(all(c(data$cat_var_new, data$time_var) %in% colnames(data$new)))
+  stopifnot(
+    is.null(data$multiplier_var) ||
+      (data$multiplier_var %in% colnames(data$new) &&
+         data$multiplier_var %in% colnames(data$old))
+  )
+
+
+  stopifnot(is.null(data$freqs_df) ||
+              (is.data.frame(data$freqs_df) && ncol(data$freqs_df) == 2))
+  stopifnot((length(unique(data$old[[data$time_var]])) == 1) &&
+              (length(unique(data$new[[data$time_var]])) == 1))
+  stopifnot(is.null(data$id_var) || ((data$id_var %in% colnames(data$old)) &&
+                                       (data$id_var %in% colnames(data$new)) &&
+                                       !anyDuplicated(data$old[[data$id_var]]) &&
+                                       !anyDuplicated(data$new[[data$id_var]])))
+}
+
+#" Validate cat2cat mappings
+#' @keywords internal
+validate_mappings <- function(mappings) {
+  stopifnot(all(vapply(mappings, Negate(is.null), logical(1))))
+  stopifnot(all(c("trans", "direction") %in% names(mappings)))
+  stopifnot(isTRUE(mappings$direction %in% c("forward", "backward")))
+  stopifnot(is.data.frame(mappings$trans) && ncol(mappings$trans) == 2)
 }